pipeline: Add predev overlay update config
Pipeline is configured to update predev deployment on merges to
both main and predev branch for now. This makes it easier for us
send PRs and test at the initial phase of the project without
sending separate PRs to predev. Once everything is on dev and prod
we can configure this to only listen for predev branch changes